keyboard interface in progress VII...

+++ src/Pointers_Container.h
@@ -2,7 +2,7 @@ template <typename IType>
 class Pointers_Container{
     public:
         string name;
-        u_int id;
+        int id;
         string default_addedItemName;
         vector<IType> itemsVector;
         vector<int> freeIdBuff;
@@ -14,8 +14,8 @@ class Pointers_Container{
         virtual IType add_itemByName(string iName, World* _world);
         IType pop(u_int index=(u_int)NULL, bool erase=true);
         IType pop(typename vector<IType>::iterator item_it, bool erase=true);
-        IType pop_itemById(u_int id, bool erase=true);
-        IType get_itemById(u_int id);
+        IType pop_itemById(int id, bool erase=true);
+        IType get_itemById(int id);
         void show_items_name_and_id();
         vector <IType> pop_itemByName(string iName, bool erase=true);
         vector <IType> clear(bool erase=true);
@@ -95,7 +95,7 @@ in "freeIdBuff" vector for future added items.
     IType item; 
     if (item_it >= itemsVector.begin() && item_it < itemsVector.end() && itemsVector.size() > 0) {
         if (erase) {
-            if ((**item_it).id < (u_int)(itemsVector.size() - 1)) {
+            if ((**item_it).id < (int)(itemsVector.size() - 1)) {
                 freeIdBuff.push_back((**item_it).id);
             }
             item = *item_it;
@@ -120,7 +120,7 @@ in "freeIdBuff" vector for future added items.
     if (index >= 0 && index < itemsVector.size() && itemsVector.size() > 0) {
         item = itemsVector[index];
         if (erase) {
-            if (item->id < (u_int)(itemsVector.size() - 1)) {
+            if (item->id < (int)(itemsVector.size() - 1)) {
                 freeIdBuff.push_back(item->id);
             }
             delete item;
@@ -144,7 +144,7 @@ void Pointers_Container<IType>::show_items_name_and_id(){
 }
 
 template <typename IType> 
-IType Pointers_Container<IType>::get_itemById(u_int id){
+IType Pointers_Container<IType>::get_itemById(int id){
     IType item;
     IType result = (int)NULL;
     typename vector<IType>::iterator item_it;
@@ -164,7 +164,7 @@ IType Pointers_Container<IType>::get_itemById(u_int id){
 }
 
 template <typename IType> 
-IType Pointers_Container<IType>::pop_itemById(u_int id, bool erase){
+IType Pointers_Container<IType>::pop_itemById(int id, bool erase){
     IType item;
     IType result = (int)NULL;
     typename vector<IType>::iterator item_it;
@@ -176,7 +176,7 @@ IType Pointers_Container<IType>::pop_itemById(u_int id, bool erase){
         item = *item_it;
         if (item->id == id){
             if (erase){
-                if (item->id < itemsVector.size() - 1) {
+                if (item->id < (int)itemsVector.size() - 1) {
                     freeIdBuff.push_back(id);
                 }
                 delete item;
